WIP: Support for native_posix

+++ src/twister2/device/fifo_handler.py
@@ -14,23 +14,26 @@ logger = logging.getLogger(__name__)
 class FifoHandler:
     """Creates FIFO file for reading and writing."""
 
-    def __init__(self, fifo: str | Path):
+    def __init__(self, fifo_in: str | Path, fifo_out: str | Path):
         """
         :param fifo: path to fifo file
         """
-        self._fifo_in = str(fifo) + '.in'
-        self._fifo_out = str(fifo) + '.out'
-        self._dump_file_path = Path(fifo).parent / HANDLER_LOG_FILE_NAME
+        self._fifo_in = fifo_in
+        self._fifo_out = fifo_out
+        self._dump_file_path = Path(fifo_in).parent / HANDLER_LOG_FILE_NAME
         self.file_in: io.BytesIO | None = None
         self.file_out: io.BytesIO | None = None
         self.dump_file: io.TextIOBase | None = None
         self._threads: list[threading.Thread] = []
+        self._created_fifos = False
 
-    @staticmethod
-    def _make_fifo_file(filename: str) -> None:
+    def _make_fifo_file(self, filename: str) -> None:
         if os.path.exists(filename):
-            os.unlink(filename)
+            logger.debug('Fifo exist: %s', filename)
+            return
+            #os.unlink(filename)
         os.mkfifo(filename)
+        self._created_fifos = True
         logger.debug('Created new fifo file: %s', filename)
 
     @property
@@ -45,14 +48,11 @@ class FifoHandler:
 
     def connect(self):
         self.dump_file = open(self._dump_file_path, 'w')
+        logger.debug('Log out: %s', self._dump_file_path)
         self._make_fifo_file(self._fifo_in)
         self._make_fifo_file(self._fifo_out)
-        self._threads = [
-            threading.Thread(target=self._open_fifo_in, daemon=True),
-            threading.Thread(target=self._open_fifo_out, daemon=True)
-        ]
-        for t in self._threads:
-            t.start()
+        self.file_in = open(self._fifo_in, 'wb', buffering=0)
+        self.file_out = open(self._fifo_out, 'rb', buffering=0)
 
     def _open_fifo_in(self):
         self.file_in = open(self._fifo_in, 'wb', buffering=0)
@@ -70,10 +70,11 @@ class FifoHandler:
             self.dump_file = None
         for t in self._threads:
             t.join(timeout=1)
-        logger.debug(f'Unlink {self._fifo_in}')
-        os.unlink(self._fifo_in)
-        logger.debug(f'Unlink {self._fifo_out}')
-        os.unlink(self._fifo_out)
+        if self._created_fifos:
+            logger.debug(f'Unlink {self._fifo_in}')
+            os.unlink(self._fifo_in)
+            logger.debug(f'Unlink {self._fifo_out}')
+            os.unlink(self._fifo_out)
 
     def read(self, __size: int | None = None) -> bytes:
         return self.file_out.read(__size)  # type: ignore[union-attr]
@@ -82,10 +83,12 @@ class FifoHandler:
         line = self.file_out.readline(__size)  # type: ignore[union-attr]
         if self.dump_file:
             self.dump_file.write(line.decode())  # type: ignore[union-attr]
-        return line
+        return str(line, 'utf-8').strip()
 
     def write(self, __buffer: bytes) -> int:
-        return self.file_in.write(__buffer)  # type: ignore[union-attr]
+        if self.file_out is not None:
+            return self.file_in.write(__buffer)  # type: ignore[union-attr]
+        return -1
 
     def flush(self):
         if self.file_in:
